+++ src/module/vue/components/impact/debilities-classic.vue
@@ -25,7 +25,7 @@
 
 <script setup lang="ts">
 import ConditionCheckbox from './impact-checkbox.vue'
-import CustomConditionCheckbox from './custom-condition-checkbox.vue'
+import CustomConditionCheckbox from './impact-checkbox-custom.vue'
 </script>
 
 <style lang="scss" scoped>
