+++ src/module/vue/components/impact/impact-checkbox-custom.vue
